Module aws_sdk_applicationdiscovery::error

Modules

authorization_error_exception

See AuthorizationErrorException

conflict_error_exception

See ConflictErrorException

home_region_not_set_exception

See HomeRegionNotSetException

invalid_parameter_exception

See InvalidParameterException

invalid_parameter_value_exception

See InvalidParameterValueException

operation_not_permitted_exception

See OperationNotPermittedException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

server_internal_error_exception

See ServerInternalErrorException

Structs

AssociateConfigurationItemsToApplicationError

AuthorizationErrorException

The AWS user account does not have permission to perform the action. Check the IAM

BatchDeleteImportDataError

ConflictErrorException

CreateApplicationError

CreateTagsError

DeleteApplicationsError

DeleteTagsError

DescribeAgentsError

DescribeConfigurationsError

DescribeContinuousExportsError

DescribeExportConfigurationsError

DescribeExportTasksError

DescribeImportTasksError

DescribeTagsError

DisassociateConfigurationItemsFromApplicationError

ExportConfigurationsError

GetDiscoverySummaryError

HomeRegionNotSetException

The home region is not set. Set the home region to continue.

InvalidParameterException

One or more parameters are not valid. Verify the parameters and try again.

InvalidParameterValueException

The value of one or more parameters are either invalid or out of range. Verify the

ListConfigurationsError

ListServerNeighborsError

OperationNotPermittedException

This operation is not permitted.

ResourceInUseException

This issue occurs when the same clientRequestToken is used with the

ResourceNotFoundException

The specified configuration ID was not located. Verify the configuration ID and try

ServerInternalErrorException

The server experienced an internal error. Try again.

StartContinuousExportError

StartDataCollectionByAgentIdsError

StartExportTaskError

StartImportTaskError

StopContinuousExportError

StopDataCollectionByAgentIdsError

UpdateApplicationError

Enums

AssociateConfigurationItemsToApplicationErrorKind

BatchDeleteImportDataErrorKind

CreateApplicationErrorKind

CreateTagsErrorKind

DeleteApplicationsErrorKind

DeleteTagsErrorKind

DescribeAgentsErrorKind

DescribeConfigurationsErrorKind

DescribeContinuousExportsErrorKind

DescribeExportConfigurationsErrorKind

DescribeExportTasksErrorKind

DescribeImportTasksErrorKind

DescribeTagsErrorKind

DisassociateConfigurationItemsFromApplicationErrorKind

ExportConfigurationsErrorKind

GetDiscoverySummaryErrorKind

ListConfigurationsErrorKind

ListServerNeighborsErrorKind

StartContinuousExportErrorKind

StartDataCollectionByAgentIdsErrorKind

StartExportTaskErrorKind

StartImportTaskErrorKind

StopContinuousExportErrorKind

StopDataCollectionByAgentIdsErrorKind

UpdateApplicationErrorKind