Module aws_sdk_applicationdiscovery::model

Modules

agent_configuration_status

See AgentConfigurationStatus

agent_info

See AgentInfo

agent_network_info

See AgentNetworkInfo

batch_delete_import_data_error

See BatchDeleteImportDataError

configuration_tag

See ConfigurationTag

continuous_export_description

See ContinuousExportDescription

customer_agent_info

See CustomerAgentInfo

customer_connector_info

See CustomerConnectorInfo

export_filter

See ExportFilter

export_info

See ExportInfo

filter

See Filter

import_task

See ImportTask

import_task_filter

See ImportTaskFilter

neighbor_connection_detail

See NeighborConnectionDetail

order_by_element

See OrderByElement

tag

See Tag

tag_filter

See TagFilter

Structs

AgentConfigurationStatus

Information about agents or connectors that were instructed to start collecting data.

AgentInfo

Information about agents or connectors associated with the user’s AWS account.

AgentNetworkInfo

Network details about the host where the agent/connector resides.

BatchDeleteImportDataError

Error messages returned for each import task that you deleted as a response for this

ConfigurationTag

Tags for a configuration item. Tags are metadata that help you categorize IT

ContinuousExportDescription

A list of continuous export descriptions.

CustomerAgentInfo

Inventory data for installed discovery agents.

CustomerConnectorInfo

Inventory data for installed discovery connectors.

ExportFilter

Used to select which agent's data is to be exported. A single agent ID may be selected

ExportInfo

Information regarding the export status of discovered data. The value is an array of

Filter

A filter that can use conditional operators.

ImportTask

An array of information related to the import task request that includes status

ImportTaskFilter

A name-values pair of elements you can use to filter the results when querying your import

NeighborConnectionDetail

Details about neighboring servers.

OrderByElement

A field and direction for ordered output.

Tag

Metadata that help you categorize IT assets.

TagFilter

The tag filter. Valid names are: tagKey, tagValue,

Enums

AgentStatus

NOTE: AgentStatus::Unknown has been renamed to ::UnknownValue.

BatchDeleteImportDataErrorCode

ConfigurationItemType

ContinuousExportStatus

DataSource

ExportDataFormat

ExportStatus

ImportStatus

ImportTaskFilterName

OrderString