Module aws_sdk_applicationdiscovery::operation

Structs

AssociateConfigurationItemsToApplication

Associates one or more configuration items with an application.

BatchDeleteImportData

Deletes one or more import tasks, each identified by their import ID. Each import task has

CreateApplication

Creates an application with the given name and description.

CreateTags

Creates one or more tags for configuration items. Tags are metadata that help you

DeleteApplications

Deletes a list of applications and their associations with configuration

DeleteTags

Deletes the association between configuration items and one or more tags. This API

DescribeAgents

Lists agents or connectors as specified by ID or other filters. All agents/connectors

DescribeConfigurations

Retrieves attributes for a list of configuration item IDs.

DescribeContinuousExports

Lists exports as specified by ID. All continuous exports associated with your user

DescribeExportConfigurations

DescribeExportTasks

Retrieve status of one or more export tasks. You can retrieve the status of up to 100

DescribeImportTasks

Returns an array of import tasks for your account, including status information, times,

DescribeTags

Retrieves a list of configuration items that have tags as specified by the key-value

DisassociateConfigurationItemsFromApplication

Disassociates one or more configuration items from an application.

ExportConfigurations

Deprecated. Use StartExportTask instead.

GetDiscoverySummary

Retrieves a short summary of discovered assets.

ListConfigurations

Retrieves a list of configuration items as specified by the value passed to the

ListServerNeighbors

Retrieves a list of servers that are one network hop away from a specified

StartContinuousExport

Start the continuous flow of agent's discovered data into Amazon Athena.

StartDataCollectionByAgentIds

Instructs the specified agents or connectors to start collecting data.

StartExportTask

Begins the export of discovered data to an S3 bucket.

StartImportTask

Starts an import task, which allows you to import details of your on-premises environment

StopContinuousExport

Stop the continuous flow of agent's discovered data into Amazon Athena.

StopDataCollectionByAgentIds

Instructs the specified agents or connectors to stop collecting data.

UpdateApplication

Updates metadata about an application.